draft: hack for presigner
